BACKPORT: cpufreq: schedutil: Add fast switch to sugov_limits path
When new cpufreq min/max limits are set, the expectation is that the CPU
frequency will be updated appropriately.  This happens for targets which do
not utilize the cpufreq fast switch APIs, but the code to do this is not
present by default for the fast switch path.  Add the necessary code to
update frequency to reflect the new cpufreq limits for targets with fast
switch enabled.

Change-Id: I211b6117005df9d340dfe0d825032cd7600cbffa
Signed-off-by: Jonathan Avila <avilaj@codeaurora.org>
Signed-off-by: starlight5234 <ifist2834@gmail.com>